Automatically generate the version number in taglib-config
git-svn-id: svn://anonsvn.kde.org/home/kde/trunk/kdesupport/taglib@1220246 283d02a7-25f6-0310-bc7c-ecb5cbfe19da

+++ taglib/asf/asffile.cpp
@@ -149,7 +149,7 @@ public:
 void ASF::File::BaseObject::parse(ASF::File *file, unsigned int size)
 {
   data.clear();
-  if (size > 24 && size <= file->length())
+  if (size > 24 && size <= (unsigned int)(file->length()))
     data = file->readBlock(size - 24);
   else
     data = ByteVector::null;
